% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_good_bad.R
\name{classify_good_bad}
\alias{classify_good_bad}
\title{Classify good / bad points}
\usage{
classify_good_bad(
  cluster_data,
  threshold = 0.55,
  nthread = future::availableCores()[[1]],
  class
)
}
\arguments{
\item{cluster_data}{\code{data.table(1)} all data points from similar tasks}

\item{threshold}{\code{numeric(1)} threshold for bad/good labeling \link{between 0 and 1}}

\item{nthread}{\code{integer(1)} number of cores to be used when training a learner}

\item{class}{\code{integer(1)} cluster identifier}
}
\value{
\code{data.table(1)} containing "good" configurations predicted with very high probability over 0.9.
All this "good" configuration will be then passed on to \code{\link{run_perf_est_model}} for building
performance estimator.
}
\description{
Function runs RandomForest Classifier with default HP config and 3 fold CV resampling
on synthetically created binary task where all configs with accuracy below 0.55 get "bad" label and
configs with accuracy over 0.55 get label "good".
}
\details{
For details refer to report in the repository.
}
