% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_results.R
\name{compare_results}
\alias{compare_results}
\title{Compare Results}
\usage{
compare_results()
}
\value{
\code{list(2)} containing default and meta performances.
}
\description{
Function should be called in the end after all results are produced.
}
