% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{default_config}
\alias{default_config}
\title{Default Configuration}
\format{
An object of class \code{list} of length 10.
}
\usage{
default_config
}
\description{
Default HyperParameter Configuration for xgboost classifier.
}
\keyword{datasets}
