% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_similar_tasks.R
\name{find_similar_tasks}
\alias{find_similar_tasks}
\title{Cosine Similarity between tasks}
\usage{
find_similar_tasks(task_id, meta_features, meta_feature_names, meta_data)
}
\arguments{
\item{task_id}{\code{integer(1)} task identifier on OpenML}

\item{meta_features}{\code{data.table()} data.table of task_meta_features}

\item{meta_feature_names}{\code{character()} meta_feature names}

\item{meta_data}{\verb{data.table(} data.table of xgboost_meta_data}
}
\value{
\code{list()} containing data_ids of 3 most similar tasks, class identifier and meta_features.
}
\description{
Using meta-features of different tasks, first mutates meta_features as described in
\code{\link{mutate_metafeatures}}
and then computes cosine similarities on them to select 3 most similar tasks for given task.
}
