% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_task_metafeatures.R
\name{get_task_metafeatures}
\alias{get_task_metafeatures}
\title{Get meta features from OpenML}
\usage{
get_task_metafeatures(task_id, meta_feature_names = meta_feature_names)
}
\arguments{
\item{task_id}{\code{integer(1)} task id}

\item{meta_feature_names}{\code{character()} names of task meta_features}
}
\value{
\code{data.table()} of task meta_features.
}
\description{
Downloads meta_features from OpenML.org for given task_id.
}
\details{
requires internet connection.
}
