% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_train_test_tasks.R
\name{get_train_test_tasks}
\alias{get_train_test_tasks}
\title{Get Train and Test tasks}
\usage{
get_train_test_tasks(task_id)
}
\arguments{
\item{task_id}{\code{integer(1)} task id}
}
\value{
\code{list()}containing train and test tasks.
}
\description{
Downloads task and its resampling structure from OpenML.
Use first train and test fold of the 10-fold CV as train, test task respectively.
}
\details{
requires internet connection.
}
