% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_model.R
\name{interpret_learner}
\alias{interpret_learner}
\title{Interpret Performance Estimator}
\usage{
interpret_learner(model, learner, task, new_data, class)
}
\arguments{
\item{learner}{(\code{Learner(1)}) mlr3 learner}

\item{task}{(\code{Task(1)}) mlr3 task}

\item{new_data}{(\code{data.table(1)}) Single test data point for Ceteris Paribus Plots}

\item{class}{(\code{character(1)}) identifier for group of similar tasks}
}
\value{
invisble \code{list()} containing plots.
}
\description{
Predictions are interpreted using simple Feature Impurity bar plots and
Ceteris Paribus Plots from DALEX package.
}
