% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{meta_feature_names}
\alias{meta_feature_names}
\title{Meta_Feature Names}
\format{
An object of class \code{character} of length 13.
}
\usage{
meta_feature_names
}
\description{
Names of the meta_features used in Analysis.
}
\keyword{datasets}
