% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_metafeatures.R
\name{mutate_metafeatures}
\alias{mutate_metafeatures}
\title{Mutate meta_features}
\usage{
mutate_metafeatures(features)
}
\arguments{
\item{features}{\code{data_table()} data.table of meta-features}
}
\value{
\code{data.table()} containing mutated meta_features
}
\description{
Mutates meta_features to make them comparable across different tasks:\cr
\tabular{cc}{
\strong{Name} \tab \strong{Fromula} \cr
MinorityMajorityClassRatio \tab \eqn{MinorityClassSize / MajorityClassSize}\cr
MaxNominalAttInstanceRatio \tab \eqn{MaxNominalAttDistinctValues / NumberOfInstances}\cr
NumberOfClasses \tab \eqn{NumberOfClasses}\cr
NumberOfFeatures\tab \eqn{NumberOfFeatures}\cr
MissingValueSizeOfFeatureMatrixRatio \tab \deqn{NumberOfMissingValues / (NumberOfFeatures * NumberOfInstances)}\cr
NumberOfInstancesWithMissingVlauesRatio \tab \deqn{NumberOfInstancesWithMissingValues / NumberOfInstances}
}
}
