% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_meta_data.R
\name{read_meta_data}
\alias{read_meta_data}
\title{Read xgboost_meta_data from csv file}
\usage{
read_meta_data(file)
}
\arguments{
\item{file}{\code{character(1)} path to xgboost_meta_data file}
}
\value{
\code{data.table()} of xgboost_meta_data.
}
\description{
Except reading csv file, function also applies log_transformation to Hyperparameters: eta, gamma, lambda, alpha.\cr
Pay attention to printed texts generated by the function, while being executed.
}
