% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_auc_distribution.R
\name{return_auc_distribution}
\alias{return_auc_distribution}
\title{Distribution of AUC}
\usage{
return_auc_distribution(data_ids, meta_data)
}
\arguments{
\item{data_ids}{\code{numeric()} numeric vector of data_ids}

\item{meta_data}{\code{data.table()} xgboost_meta_data}
}
\value{
\code{table()} of frequencies of AUC.
}
\description{
Return table of distribution of AUC for given data_ids.
}
