% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_config.R
\name{run_config}
\alias{run_config}
\title{Run XGBoost Classifier}
\usage{
run_config(
  task_id,
  config = NULL,
  meta_config = TRUE,
  default = FALSE,
  seed = NULL,
  nthread = future::availableCores()[[1]],
  meta_features
)
}
\arguments{
\item{task_id}{\code{integer(1)} task id}

\item{config}{\code{list()} named list of configurations in case neither default nor meta_config, but custom config}

\item{meta_config}{\code{logical(1)} if TRUE meta-analysis is done to find configuration}

\item{default}{\code{logical(1)} if TRUE learner is trained with default configuration}

\item{seed}{\code{numeric(1)} seed for reproducibility}

\item{nthread}{\code{integer(1)} number of cores to be used when training and predicting with learner}

\item{meta_features}{\code{data.table()} task meta_features}
}
\value{
invisible \code{list()} containing resulting AUC and train_time measures.
}
\description{
Function has 2 modes:\cr
\enumerate{
\item If default is set to TRUE, XGBoost is trained with default configuration.
\item If meta_config is st to TRUE \code{\link{metalearn_config}} gets called to start
meta-learning pipeline and after finding it XGBoost is trained with meta configuration.
}
}
\details{
requires internet connection.
}
