% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_perf_est_model.R
\name{run_perf_est_model}
\alias{run_perf_est_model}
\title{Predict performance}
\usage{
run_perf_est_model(
  perf_est_data,
  nthread = future::availableCores()[[1]],
  class
)
}
\arguments{
\item{perf_est_data}{(\code{data.table(1)}) meta_data}

\item{nthread}{(\code{integer(1)}) number of cores to be used when training a predicting learner}

\item{class}{(\code{integer(1)}) identifier of the group of similar tasks}
}
\value{
\code{data.table()} of performance predictions on the 5 test sets.
}
\description{
Runs RandomForest Regression with default HP config and 5 fold CV resampling
on the "good" configurations resulted by the function call \code{\link{classify_good_bad}}.
}
