% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_configs.R
\name{select_configs}
\alias{select_configs}
\title{Return configuration with highest predicted accuracy and fastest training time}
\usage{
select_configs(perf_predictions, perf_est_data, sigma = 0.005)
}
\arguments{
\item{perf_predictions}{\code{data.table()} data.table of performance predictions}

\item{perf_est_data}{\code{data.table()} data.table of data used in training performance estimator}

\item{sigma}{\code{numeric(1)} parameter for accuracy range in fastest configuration decision}
}
\value{
\code{list()} containing configs with highest predicted accuracy  and configs with fastest train time for each similar task.
}
\description{
After estimating performace predictions for similar tasks, for each similar task select:
\itemize{
\item One config with highest predicted accuracy
\item One config with smallest expected train_time, whereas accuracy is larger than \eqn{HighestPredAUC - sigma}
}
}
