% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{test_ids}
\alias{test_ids}
\title{Test IDs}
\format{
An object of class \code{numeric} of length 20.
}
\usage{
test_ids
}
\description{
test_ids refers to task_ids from OpenML.org used in Analysis.
}
\keyword{datasets}
